refactor(storage): recursive calls to MMKV storage object fixed
