+++ openwink-app/Storage/Storage.ts
@@ -1,0 +1,2 @@
+import { MMKV } from "react-native-mmkv";
+export default new MMKV();
