+++ openwink-app/Storage/index.ts
@@ -1,8 +1,3 @@
-import { MMKV } from "react-native-mmkv";
-
-export default new MMKV();
-
-
 export * from "./AutoConnectStore";
 export * from "./CustomCommandStore";
 export * from "./CustomOEMButtonStore";
